Move timestamp to own file
